All calculations now update on the json.

+++ Cobit-19/Shared/Dtos/FullObjectiveAuditDto.cs
@@ -6,12 +6,12 @@
         public string objectiveDescription { get; set; }
         public string objectivePurpose { get; set; }
         public int maturityLevel { get; set; }
-        public float maturityLevel0PercFinal { get; set; }
-        public float maturityLevel1PercFinal { get; set; }
-        public float maturityLevel2PercFinal { get; set; }
-        public float maturityLevel3PercFinal { get; set; }
-        public float maturityLevel4PercFinal { get; set; }
-        public float maturityLevel5PercFinal { get; set; }
+        public double maturityLevel0PercFinal { get; set; }
+        public double maturityLevel1PercFinal { get; set; }
+        public double maturityLevel2PercFinal { get; set; }
+        public double maturityLevel3PercFinal { get; set; }
+        public double maturityLevel4PercFinal { get; set; }
+        public double maturityLevel5PercFinal { get; set; }
         public List<ComponentDto> components { get; set; }
     }
 }
